add geocode and other stuff
