+++ app/assets/javascripts/application.js
@@ -1,7 +1,2 @@
 //= require rails-ujs
 //= require_tree .
-import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the stylesheet_pack_tag in the layout!
-
-import { initMapbox } from '../plugins/init_mapbox';
-
-initMapbox();
