--- app/assets/javascripts/channels/plugins/init_mapbox.js
@@ -1,15 +1,0 @@
-import mapboxgl from 'mapbox-gl';
-
-const initMapbox = () => {
-  const mapElement = document.getElementById('map');
-
-  if (mapElement) { // only build a map if there's a div#map to inject into
-    mapboxgl.accessToken = mapElement.dataset.mapboxApiKey;
-    const map = new mapboxgl.Map({
-      container: 'map',
-      style: 'mapbox://styles/mapbox/streets-v10'
-    });
-  }
-};
-
-export { initMapbox };
