+++ app/javascript/packs/application.js
@@ -1,1 +1,6 @@
 import "bootstrap";
+import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the stylesheet_pack_tag in the layout!
+
+import { initMapbox } from '../plugins/init_mapbox';
+
+initMapbox();
