+++ app/views/kitties/index.html.erb
@@ -4,23 +4,31 @@
 
 
 
-  <tbody>
-    <% @kitties.each do |kitty| %>
-      <tr>
-        <td><%= kitty.first_name %></td>
-        <td><%= link_to 'Show', kitty %></td>
-        <% if policy(kitty).edit? %>
-          <td><%= link_to 'Edit', edit_kitty_path(kitty) %></td>
-        <% end %>
+<tbody>
+  <% @kitties.each do |kitty| %>
+    <tr>
+      <td><%= kitty.first_name %></td>
+      <td><%= link_to 'Show', kitty %></td>
+      <% if policy(kitty).edit? %>
+        <td><%= link_to 'Edit', edit_kitty_path(kitty) %></td>
+      <% end %>
 
-        <% if policy(kitty).destroy? %>
-          <td><%= link_to 'Destroy', kitty, method: :delete, data: { confirm: 'Are you sure you want to delete this lovely kitty ???' } %></td>
-        <% end %>
+      <% if policy(kitty).destroy? %>
+        <td><%= link_to 'Destroy', kitty, method: :delete, data: { confirm: 'Are you sure you want to delete this lovely kitty ???' } %></td>
+      <% end %>
 
-      </tr>
-    <% end %>
-  </tbody>
+    </tr>
+  <% end %>
+</tbody>
 
+
+<div
+  id="map"
+  style="width: 100%;
+  height: 500px;"
+  data-markers="<%= @markers.to_json %>"
+  data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"
+></div>
 
 <br>
 <br>
