+++ app/views/kitties/show.html.erb
@@ -27,13 +27,7 @@
   </div>
 </div>
 
-<div
-  id="map"
-  style="width: 100%;
-  height: 600px;"
-  data-markers="<%= @markers.to_json %>"
-  data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"
-></div>
+
 
 <% if policy(@kitty).edit? %>
   <%= link_to 'Edit the KITTY', edit_kitty_path(@kitty) %> |
